services/payment-service/requirements-minimal.txt
# Core FastAPI dependencies (no Stripe)
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
aiohttp==3.11.10
python-multipart==0.0.12
python-dotenv==1.0.1
alembic==1.13.3
requests==2.32.3